Projects/4_NewRAG/requirements.txt
langchain
langchain-community
langchain-groq
streamlit
chromadb
python-dotenv
wikipedia
sentence_transformers
pypdf
